+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 * Increased the max number of logs from 4 to 10 included in issue reports.
 * Changed VersionedFile to use temp file with process ID to avoid collisions.
 * Display tooltips for Settings widget labels, not just values.
+* Fixed device selection in Multimeter, Value, and Accumulator widgets.  #233
 
 
 ## 1.0.42